feat/Showing and displaying posts and comments

+++ src/main/java/rs/ac/uns/ftn/informatika/rest/repository/IPostRepository.java
@@ -3,7 +3,7 @@ package rs.ac.uns.ftn.informatika.rest.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import rs.ac.uns.ftn.informatika.rest.domain.Post;
 
-public interface IPostRepository extends JpaRepository<Post, Long> {
-    public Post deletePostById(Long id);
+public interface IPostRepository extends JpaRepository<Post, Integer> {
+    public Post deletePostById(Integer id);
 
 }
